Update backend based on code review
